CentralAuthUser: Don't load user information unless needed
When checking if a user can log in, only load the local user existence
data if the user is suppressed and so the information would be used. The
CA data is always loaded at that point, and the number of users who
would hit that code path (suppressed but not locked) is very small.

Also remove the outdated FIXME.

Change-Id: I8d3666d8253fb77a0be994139f91265a7e0981a5