GlobalRenameUser: Ensure old username is in canonical form
Apparently CentralAuthUser does not do username normalization. (I filed
T343963 about that). So when bb89a133 ("Add a GlobalRenameFactory")
changed Special:GlobalRenameUser to use
CentralAuthUser::getInstanceByName instead of round-tripping the
username via an User object, the queryAttached() call in
GlobalRenameUser started returning results that don't make sense. This
patch fixes that and adds a check to GlobalRenameFactory to throw an
exception if that happens again.

Bug: T343958
Change-Id: I682082d4c97a39001af6ef2fc5153a2759892e4a