Don't delete central session if it is rejected for local reasons
If the central user is a temp user, and the user visits a wiki in the
same cookie domain which does not have temp users enabled, the user name
will not be usable and the central session cannot be used. But if the
user returns to the original wiki, we want the central session to still
exist. CentralAuthSessionProvider is eager to delete shared cookies at
the first sign of trouble.

The same arguments apply to existing but unattached users, and this
issue probably affected them back when that was a thing.

So:

* When delivering local session cookies, don't delete the central
  cookies.
* In the case of an unusable name or unattached account, if there is no
  local session, create a new one and mark it "persisted". Otherwise the
  session will be unpersisted and central cookies will be deleted.
* On auto-login in the setCookies stage, don't reset the session ID if
  the name is unusable or has a local conflict.

Also:

* In the auto-login setCookies stage, factor out the duplicate
  User::newFromName() calls. We don't need a UserIdentityLookup because
  we already have a User object, and a User is a UserIdentity. We could
  use a UserFactory but that is a job for another commit.

Bug: T342475
Change-Id: I5fe0c5a64770f5f41007bcbf9cfa253eed13a979