SpecialGlobalGroupMembership: Normalize usernames
CentralAuthUser does not do username normalization
(see T343963). With fcece057670, a code path that
used to do the normalization for SpecialGlobalGroupMembership
was removed and consequently, Special:GlobalGroupMembership
stopped working for non-canonical username forms.

Wrap $username in UserNameUtils::getCanonical()
in SpecialGlobalGroupMembership::fetchUser(),
to ensure the global user is fetched using
the canonical username.

Change SpecialGlobalGroupMembershipTest to test
against a non-canonical username to catch those
regressions in the future.

Bug: T344495
Change-Id: Ia81f752669e12d430c2c8d0b95b0c297025842f0