Remove success notification when we can't get autologin response
The comment on this code says that "This happens if the user is
logged in securely, while also auto-loggedin from an http page.".
This is a very optimistic view: it happens whenever the autologin
request fails, or it succeeds but we can't access the response
due to browser security reasons (CORS, etc.).

Since we stopped supporting HTTP on Wikimedia wikis many years ago,
the optimistic case can never happen, while various problematic
cases happen often (see the bugs below).

Bug: T56350
Bug: T103883
Bug: T307337
Bug: T348852
Change-Id: Ic110483ed792ff1e17f014f6268d8a848dd56382