Special: Fix potential double invokation of PostLoginRedirectHook
This patch attempts to fix a double invokation of PostLoginRedirect
hook on successful logins (same web request) if an extension makes
use of the hook in it's code.

SpecialCentralLogin already provides a hook for doing redirects, see:
CentralAuthPostLoginRedirect (which does exactly that) and if CA is
enabled, this hook should be used instead for perform redirects.

On wikis without CentralAuth, the PostLoginRedirectHook can be used
to perform redirects after successful logins.

This patch will enable MobileFrontend for example that uses the
UserLoginCompleteHook incorrectly (as of today) to make use of the
PostLoginRedirectHook without calling the hook twice.

Bug: T350138
Change-Id: Ie6e3fa24f4881855bd41c0361b8320c7f34918de