Have Special:GlobalGroupPermissions display user group technical name
Old behavior is to display the human readable name (e.g. Edit
filter helpers), "management" link, and a "list of members"
link. New behavior is the same, except it also displays the
technical name (e.g. abusefilter-helper) enclosed in <code>
tags.

This is in a similar style to the rights that are displayed in
the adjacent column. This also syncs this behavior up with
Special:UserGroupRights.

Bug: T351612
Change-Id: I3db7ff1db3c1b6afaf1e52bf6271e3944e0f8e78