Sort global user groups in CentralAuthUser::loadGroups
Sort global user groups in backend to ensure the order of the groups and
the order in the argument 'conflictcheck-originalgroups' on
SpecialGlobalGroupMembership are in the same order and matching.

Use a sort in php as the query is complex with joins and an extra order
by could result in performance issues. The list of user groups is not
that big, so do a ksort in php.

The test under postgres gets a "userrights-conflict" failure,
the test failure shows that the expected change is not done.

1) MediaWiki\Extension\CentralAuth\Tests\Phpunit\Integration\Special\SpecialGlobalGroupMembershipTest::testSave
Failed asserting that two arrays are equal.
--- Expected
+++ Actual
@@ @@
 Array (
-    0 => 'group-one'
+    0 => 'group-two'
     1 => 'group-three'
 )

tests/phpunit/integration/Special/SpecialGlobalGroupMembershipTest.php:292

Change-Id: I47b311cdcf3618b1ee543f6ea9cde5246c598624