Add abstract schema change for global_user_autocreate_serial.uas_year
The new uas_year column is needed so that the index for the unique
part of a temporary username can start again each year. It is used
when the following configuration are used:

$wgAutoCreateTempUser['serialProvider']['type'] = 'central';
$wgAutoCreateTempUser['serialProvider']['useYear'] = true;

The new column is added to the local table in
If51acb3f4efa361ce36d919c862a52501a5a7d24.

Note that schema change patches for CentralAuth are not run via
update.php and must be run manually.

Bug: T349501
Change-Id: I6b3c640a4e74f52fd4a4f46de5a2cbe80fe3b665