CentralAuthSessionManager: Use MicroStash service for CA tokens
wgCentralAuthTokenCacheType is now obsolete with the presence of
wgMicroStashType which is wrapped via the micro stash service making
use of `$wgObjectCaches['mcrouter-primary-dc']`.

Test plan
=========

MicroStash under the hood consumes mcrouter-primary-dc which is the
same configuration that wgCentralAuthTokenCacheType used, so technically,
there should be no change at all on this front, everything should just
behave the same.

The behavior that this patch changes (in code) which may possibley affect
third-parties if they use CentralAuth is the fallback logic when token cache
type is not configured (as it was before this change). Now we're always
using MicroStash for the purpose of storing our tokens (T278392).

Once this change goes live, let's make sure CentralAuth tokens still
work as expected.

Bug: T336004
Change-Id: Id67dd87fe3b18cf8f0c02112de8ec61d971c1844