User: Normalize username of user if in non-canonical form
Sometimes, callers of `getInstanceByName()` can pass in the username
as a string in a non-canonical form and this will break SQL queries.
In order to avoid this, let this method normalize the username supplied
for correctness and stability of the system.

Also, if a username is external, CA will ignore it completely with this
patch and not throw.

NOTE: In the case where `$username` is an IP address, make sure we validate
it and santize it but we can't canonicalize it because it's an IP, we
will only do that for actual usernames.

Bug: T343963
Depends-On: I11937d0d53b3914d7f6bd1483404a04b74ac26c9
Change-Id: I2c7b4ceb56463e881305287a030448c9aa288532