CentralAuthUser: Fix $mGroups containing duplicate or missing items
We added each group to $mGroups as many times as many rights were
associated with it, because the data was pulled out from the query for
the rights. Follow-up to d09aefd0b6dff53499e614a4d91cc3283881ccb6.

Duplicate items were filtered out elsewhere, so they had no visible
impact, it was just inefficient; but if a group had no rights
associated with it, it wouldn't be listed, which was a bug.

Also remove unnecessary query for the WikiSets used by the user's
groups, which we were not using (and document why we can't use it).

Change-Id: I99495f0e5da53ee29c6d4cd5a102e553ee58fcb7