SpecialGlobalGroupMembership: Fix adding user to groups they're already in
The condition in array_filter() was always true, because it was trying
to check whether a group name is missing in an array that countains
group expiry values. It needs to check a key instead of value.
This didn't cause user-visible issues, because addToGlobalGroup()
uses Database::replace(), thus being able to handle existing groups.

I found this while trying to figure out the Phan suppression, and
apparently fixing that mistake also clears the Phan problem.

Change-Id: Ifa271e68457b23cc329315696d90789fd258d204