Don't parse HTML as wikitext on Special:GlobalGroupPermissions
Same as I74fcdab in core.

These pieces are known to be trivial HTML. The content is known to
be a plain text identifier. We don't need to parse it again as if it
is wikitext. rawParams() is the way to mark such parameters as "show
as is, don't parse again".

The double escaping was added as part of I90f39be in 2017. It was
most certainly never doing anything because the incoming identifiers
are plain ASCII with nothing to escape.

Change-Id: Id68f3f519f2f456dd6ecd7424efbde10a290c310