Use a factory to construct GlobalRenameUserStatus objects
Update the GlobalRenameUserStatus class to use a factory constructor
class like the GlobalRenameUser class already does. This means we can
inject parameters into that class cleanly. All uses in CentralAuth are
updated, I will submit a follow-up for the single use in
WikimediaMaintenance.

The static method getInProgressRenames() is untouched, for now,
although a future patch could move it to GlobalRenameFactory or some
other class where the database can be injected.

Change-Id: I693ced583132ae4e2a9890a0a971b6a14d453c9c