SpecialGlobalGroupMembership: Avoid deprecated Xml::checkLabel
This method has a ton of layers of indirection, which makes for an
unintuitive signature, as well as various compromises in its API
design that are not something you can reasonably remember and learn.

```
Xml::checkLabel( $member, "wpGroup-" . $group,
    "wpGroup-" . $group, $set, $attr );
```

You can stare at this for quite a while and still have no confidence
in what it actually produces. You have to either run it in a REPL,
or setup up an elaborate UX test to find the exact scenario in a local
environment to inspect the DOM. The method docs don't mitigate this,
either, but with such a bad signature, docs would indeed be at best
help mitigate and manage expectations, instead of actually making it
intuitiive, productive, or memorable.

Xml::checkLabel() takes 4 unnamed parameters, of which two are often
the same value passed twice (name and ID), one applies only to the
label ($labelText) which is actually the first parameter despite being
used for only the last element in the output. It produces the `<input>`
element first and the `<label>` after it. The `$attrs` surprisingly
applies to both elements but in different ways. They're all blindly
copied to the input element, but then some hardcodes subset gets
duplicated to the `<label>` element as well.

> Xml::checkLabel(1, 2, 3, 4) modifies $attrs with ID.
-> Xml::check modifies $attrs again with $name, but ignores 1=labelText
-> -> Xml::attrib
-> -> Xml::element
-> Xml::label ignores most but not all (!) of the 4=attrs parameter.
-> -> Xml::element

Replace it all with a simple Html::element() instead.

Change-Id: I61c8f6712734dd07ce6f7b8d14f7d88de671aab4