[SUL3] Enable users to login on Wikimedia's central login wiki
This patch changes how the authentication work flow happens.
Previously, authentication will happen on the local/initiating
wiki (whatever wiki the user visits) but with this patch, all
login happens on the central login domain.

CentralAuth will check if the wiki that the user is about to
perform the action on is the central login wiki or not and if
the former is true, things will proceed normally, otherwise,
they'll be redirected to the central domain and to perform the
authentication there (step 1), gets redirected back to the local
wiki for authentication to complete (step 2).

The patch does the following:
1. Uses the CA primary authentication providers for the new SUL3
   login flow.
2. When favourable conditions are set, redirect to the central wiki
   to perform the login action there and then continue in the local
   wiki.
3. Forward a `returnurl` to the central login wiki, so it knows
   where to redirect back upon successful central login.
4. Provides a basic Redirecting provider which displays a button
   when the provider is registered.
5. Use query parameters to determine which wiki and corresponding
   page to redirect back to after a successful login.
6. Once the user is centrally logged in, the user is redirected back
   to the home wiki and then AuthManager tries to continue the login
   on the local wiki from there. If this succeeds, the user gets logged
   in locally based on the central wiki providing it a username token.
7. Implemented the ?usesul3=1 query flag to trigger the login form on
   the central login wiki with all the relevant parameters forwarded.
   This way, we can test the normal login (SUL2) and the new SUL3 flow
   side-by-side.

Bug: T363483
Change-Id: I078925a5e9979b0f7aa56387a07d20368c5b7fff