Remove redundant constructor tests
These tests don't really do anything. The only way they could ever
fail is when the constructor throws an exception. But this isn't
even possible as there is no code that could do that. Even if such
code is added in the future, this is still covered by all the other
tests that also call the constructor.

Change-Id: I798943b770c58b5ab9e655fd731e2d5600c10789