Add SSO domain handling
Declare the $wgCentralAuthSsoUrlPrefix global, use it to
detect whether the current request is to the SSO domain,
and apply various limitations when it is.

Soft-depends on I4a120d3024f655f8a1f020d980373ebd56b423bc.

Since we aren't sure yet whether we'll use the shared domain
mechanism, the patch focuses on functionality over
maintainability (e.g. extensions should be able to declare
their own API endpoints as necessary for authentication).
If we decide to keep this code, that should be fixed in a
followup.

Bug: T365162
Change-Id: I05be0088a262d902d86c86300e4068ac943ad6a6