Add rq_type column to renameuser_queue
This new column is used for tracking the type of rename request being
made. Account vanishing will make use of existing rename schema and code
as much as possible. A value of 0 being stored in rq_type indicates a
typical rename request (and is default), while 1 indicates a vanish
request.

Bug: T366440
Change-Id: I1fc0c59339c7155d4ca797bf6cb97a3c317df74a