[SUL3] Tokenize sensitive URL query param for security
In order to avoid URL parameter pollution, parameter tampering or
certain fixation attacks, let's not expose the return URL & username
transparently (as plain text) in the URL and reveal tokens while we
keep the actual data in the token store. When needed, we'll use the
tokens in the URL query param.

During the login-continue authentication action in the local wiki,
we'll read the value of the token from the URL and then check if
these tokens have entries in the token store with valid data and use
the data.

Bug: T369650
Change-Id: I0fa64fe03549726b434fcb4747fbf878ab51f5d4