CentralAuthUtilityService: Avoid use of deprecated `clearLastError()`
BagOStuff::watchErrors() returns a token that can be used with the
BagOStuff::getLastError() method to get the last error registry. Use
this instead of deprecated clearLastError().

This new style for clearing last errors was introduced in 2021,
I281817a85602967c0ec2bdd23a5d8be101680b64 (I281817a85602967) as it
is more robust due to its support for allowing nested callers.

An outside BoS instance might want to watch the errors returned by
multiple internal BoS instances and with the use of watch point
tokens, this is possible.

Change-Id: I0015ac96dcbe4b2f1e5b82b50c7e67f04052206b