Improve error handling in batchVanishUsers script.
...specifically when creating and writing the output.csv file.
There can be cases when fopen() returns successfully, but the file
is still not writable due to permissions. In such a case, the script
will proceed to do all of its work, but none of the output will be
captured.

This will update the script to fail, prior to doing anything else,
if it detects that the output file is not writable.

Bug: T369489
Change-Id: I8519f9eacf522db17bd1a9b5383bf0eede08a5ae