Use TokenManager for existing token operations
Use tokenize()/detokenize() for central login and central
autologin. In theory this is a no-op - functional changes
are left to a follow-up patch to minimize risk.

* Relax the type of the $value parameter in tokenize()
  so one can tokenize arrays of data, not just strings.
* Allow the key parameter in tokenize()/detokenize() to be
  an array to represent composite keys. These will be applied
  in a weird order because the existing code puts part of
  the namespace before and part of the namespace after the
  token; it did not seem worth the effort of a key migration
  to fix that.
* Add an options parameter and some alternative methods to
  account for small variations in how tokens are used.
* Replace all existing uses of the token store, and make the
  token store an internal detail of the token manager.

Also make some small code style improvements in touched code.

In practice, there seems to be one small behavior change:
when using an invalid CA token for an action that should
trigger temp user creation, before the patch it resulted in
temp user creation, with the patch it results about an
(unhelpful) error about not being able to create the temp
user.
Not sure what exactly is happening there, but it's an
improvement, sort of; and in any case this is a shortcoming
of the REST API that should be fixed in a more generic way
(T252591).

Bug: T363699
Change-Id: I560551f49aae88daf993d62c2f089223fe549c02