SUL3: Allow using "remember me" set on the shared domain
In this patch, we enable user who are performing the login action
in SUL3 mode to be able to transfer the "remember me" state on the
central domain to the local domain.

So if the user checks the "remember me" box in the central domain's
login form, that will be applied to the local wiki automatically.
This is achieved by using the centralauthLoginToken URL query param
data upon detokenization to read the remember me flag there an apply
on the authentication manager accordingly.

Test plan
=========

Perform authentication with SUL3 mode enabled and ensure that the
user's session cookie has been set to expire for an extended period
of (the default) is 180 days instead of 24 hours for normal logins.

For normal logins, verify that the user's central and local session
cookies are set to expire in 24 hours (or whatever is configured).

For "remember me" logins, verify that the user's central and local
session cookies are set to expire for an extended period (default is
180 days) and also check that the `centralauth_Token` cookie is set.

Bug: T369668
Depends-On: I6e2e2d892d2b777cb4757c7c0d222afad9da506c
Change-Id: I8862ed18ef458d39eb365529110cfc62e6871461