SUL3: Expand ResourceLoader URLs
Use the ResourceLoaderModifyStartupSourceUrls hook to ensure that
the URLs that appear in the the startup module JS code are
absolute (so that when we are on the SSO domain and have set
$wgLoadScript to use the canonical domain, script URLs generated
by JS based on the code in the startup module also use the
canonical domain.

When MobileFrontend is used, also ensure that the script URLs use
the mobile domain if the embedding page was using it.

Bug: T365162
Bug: T371530
Depends-On: I199ab779abd0596b836ae43dcc5f2f2a489c9274
Change-Id: I635cdaf8a0a48b073e78604af641643b8be90c77