AuthenticationProvider: Remove unnecessary isSharedDomain() checks
The checks in the two providers don't need to be mutually exclusive;
it's harmless to check user existence twice (and it's cached).

The old code was correct as long as both providers were enabled on the
wiki, but if we ever configure them so that only one is enabled, using
these conditions to skip the checks could become incorrect.

Change-Id: I2e5073ebb108128903d449a9fdd7d61f7f4fdf3c