SUL3: Use a session secret to prevent session fixation attacks
Replace the tokenized return URL and username with more generic
data, and include a secret in the data so it gets roundtripped
and can be used to verify that we are in the same session we
started from. Rename related constants and query parameters.

Add central user ID to the data when returning to the local
wiki, to make sure we never log someone into an unrelated account
if somehow not all users are attached to the identically-named
central account.

Bug: T363699
Change-Id: Ic28080238b37f1772efc625b367e9c52e74eb7b5