SUL3: Support signup/account creation action in SUL3 mode
In this patch, when a user clicks on the "Create Account" link on
the local wiki, CentralAuth should detect if we're in SUL3 mode and
redirect the action initiator to a shared domain for them to create
their user account there (via a URL query parameter 'sul3-action').

This flows is captured in the following matter:
-> visit: local wiki
-> click: local wiki CreateAccount
-> redirect: local wiki UserLogin (param 'sul3-action')
-> redirect: shared domain CreateAccount
-> action: fill CreateAccount form on shared domain
-> redirect: local wiki UserLogin
-> action: user is logged in (on success)

An 'sul3-action' query parameter is used to direct what the primary
authentication provider should do on the shared domain. When it sees
the 'signup' action, it triggers account creation instead of the default
login.

NOTE: SharedDomainUtils has abstracted a method for treating various
actions so that in the future, if we have other actions different from
'signup' or 'login', we can generate a URL there specific to the action
on the shared domain and terat it.

Test plan:
==========

Follow the steps listed above and verify that your account was created
successfully and you're logged into the local wiki.

Bug: T370813
Change-Id: I02a76c8b8d0dee0f66d98039f2595ba1338b3290