Don't show "SUL finalization" message if reason field is empty.
When the user-initiated "reason" for vanishing is empty, the rename
request page displays a fallback message related to "SUL finalization".
This was added before my time, so I'm not sure of the exact reason for
showing this fallback message, but I'm pretty certain that SUL has
indeed been finalized by now, so this is safe to remove.

Bug: T372223
Change-Id: Iecfa06b113c8dd55c866d9f2898a9a640fc6393c