Fix "Keep me logged in" with $wgCentralAuthCookies = false;
Does not affect WMF production, because we set that config to true.

Ever since I installed CentralAuth on my local test wiki, the "Keep me
logged in" checkbox was not working – I would become logged out after
just a few hours. It turns out that CentralAuthSessionProvider would
always disable the core "remember-me" handling, but only enable its
own when $wgCentralAuthCookies was set to true (the default is false).
Now it keeps the core handling enabled when not enabling its own.

Change-Id: Ia8ca12941eca5b9b1dbf200139933d787e842599