SUL3: Forward `campaign` parameter on account creation in SUL3 mode
This patch makes it possible to forward the `campaign` query parameter
in the case that a user is visiting the account creation page through
a link with a campaign identifier.

This parameter forward will let the campaign logic to track users coming
from such a campaign via the local wiki to the shared domain.

Bug: T370810
Change-Id: I50ee7550b879dfc45a44a70ff3f621a0b2137083