SUL3: Inherit desktop view to shared domain for uniform experience
On normally circumstances, if we're moving forward with an SSO domain,
this is not supposed to happen except the shared domain is a real wiki.

Let's protect against the case where our shared domain is in a mobile
view state and the local domain that the user is coming from is in a
desktop view state. In such a case, let's inherit whatever the local
domain view is in and transfer that experience to the central domain.

The patch: I3ac9cee66ae72b97d7d46f9b6159741cd4d3515b nicely handles
the mobile view state inheritance across local to shared but doesn't
handle the opposite case. That is handled in this patch.

Change-Id: I92d611848797ff82cd64ae01bce551fd4ec10a0d