SharedDomainUtils: Add cookie-based SUL3 activation mechanism
This patch enables a cookie-based SUL3 activation mechanism to
facilitate testing the SUL3 feature. A user will be required to
manually create this cookie on their browser to trigger this option
to be enabled.

In addition to other ways of triggering SUL3 mode like 'always' and
'query-flag', just like `query-flag`, this option is fully controlled
by the user and is sticky (on browser side).

If the user wants to disable this cookie-based flag, they should
delete the cookie.

Create a cookie using devtools on your browser with the name:
`sul3OptIn` and value: `1`, then for example try to login. You
should see that SUL3 mode is enabled and login should happen on the
shared domain.

Bug: T375787
Change-Id: I69cda7cf73aa12c54d4f365bdb153f5ff273cd81