UserSelectQueryBuilder: temp() should not select non-temp users
When temp users are not configured, temp() was a no-op; it should
be the empty condition instead.

Didn't cause any problems (the only caller is
expireTemporaryAccounts.php and that does check the isKnown() flag
first) but would be an easy mistake to make in the future.

Corresponding core patch: I904bbb97b028702ecfe98c97e3f821ba6038e27a

Change-Id: I107222c77d8f6dc4caee0db4f706108a2581b4bd
Bug: T372702