Special: Add support for central autologin in SUL3 mode
This patch adds support for performing central autologin in SUL3
mode while maintaining the same behavior in SUL2 mode. When a user
successfully authenticates via the shared domain (in SUL3), all the
wikis listed in the autologin wikis config map should also attempt
an automatic login in the background (edge login) so that by the time
you visit any of the autologin wikis, you're automatically logged in.

NOTE: This patch also contains some minor refactoring and cosmetic
changes like improvement of docs to better clarify the responsibility
of `getUrl()` in CentralDomainUtils class and `getUrlForSharedDomainAction()`
in SharedDomainUtils. They're different because one is in SUL3 mode
only while the other is more generic and can be used in SUL2 & 3.

As an addendum, in SUL3 mode, the shared domain is not a real wiki,
so, the logic relies on performing operations on the URL of the shared
domain which is indicated by fake/pseudo wiki ID (#central#) to be
used in certain places like `CentralDomainUtils::getUrl()`. The pseudo
wiki ID is not to be used with WikiMap class as it is not a real wiki.

Test plan
=========
Locally, verify that the following still work:

* In SUL2 mode, central autologin and central login still works like
  today.
* In SUL3 mode, all the wikis in the auto login wikis config map should
  also have a valid logged-in session or at least trigger a login flow
  via invisible pixels embedded in these wikis.
* In SUL3 or SUL2 mode, the behavior of the central autologin experience
  stays the same.
* Authentication actions (login/signup) on the mobile domain keeps the
  experience and doesn't unexpectedly switch to desktop when supposed
  to mobile and vise versa.

In SUL2 Mode:
- Visit a local wiki in your local wiki farm
- Login on the local wiki and visit other wikis in the wiki farm
- Verify that you're logged-in in all those wikis.

In SUL3 Mode:
- Visit a local wiki in your local wiki farm
- Try to login and see that you get redirected to the shared domain
- Login on the shared domain & get redirected back to the local wiki you
  came from
- Verify that you're also logged-in in all the rest of the wikis in your
  autologin wikis map.

In SUL2/3 mode on mobile device/mode:
- Visit a local wiki in your local wiki farm (in mobile mode)
- Click the "Login/CreateAccount" link on the local wiki
- (For SUL3, you'll be redirected to the login/create account page in
  mobile mode on the shared domain wiki. Meaning the shared domain should
  inherit the mobile view from the local domain.)
- Login on the shared domain & you'll get redirected back to the local
  wiki (still you should be in mobile mode)
- Verify that you're also logged-in in all the rest of the wikis in your
  autologin wikis map.
- Verify that in SUL3 mode, the experience stayed on the device specific
  views and in SUL2 mode, you redirect back to the local wiki and still
  are on a mobile experience view.

Bug: T376488
Bug: T375788
Change-Id: I7d86c7dc697cd18e96b53b2058e60edad2bec77b