Fix warning icon color in preferences
`'flags' => [ 'destructive' ]` did not work correctly, because the
'alert' icon doesn't have a 'destructive' variant defined in OOUI.
Instead the icon was always black, even in dark mode.

I think at this point it would be unexpected to fix that bug
(and make the icon red), so just remove that, allowing the icon
to use the default color (black by default, white in dark mode).

Bug: T377660
Change-Id: I25103b90cf50e38588fb6cc892106c2bdc0f24ad