SUL3: Disable SUL3 mode with `usesul3=0` query flag
This patch adds support for disabling SUL3 mode using the inverse
of the `usesul3=0` parameter. Setting the named parameter to value
"1" enables SUL3 but setting it to "0" should disable it to use the
current SUL2 mode.

Disabling SUL3 mode via the query parameter with this patch is now
comes in 2 ways:

1. Set the query parameter `usesul3=0` in the request URL.
2. Ignore and don't set the query param at all in the request URL.

Notice that the precedence has been reversed. Before this patch, the
precedence was: `always` > `cookie` > `query-flag`, but with this
patch, it is: `query-flag` > `cookie` > `always`.

So that mean, if we have a situation like: `[ 'always', 'query-flag' ]`,
and we send a URL with `&usesul3=0`, SUL3 will be disabled.

Test plan
=========
In your LocalSettings.php:
* set `$wgCentralAuthEnableSul3 = [ 'always', 'query-flag' ];`
* set $wgCentralAuthSsoUrlPrefix = "<URL to SSO domain>";

Then visit:
* the login page of a local wiki your local setup and add `&usesul3=1`,
  that should redirect you to the SSO domain.
* visit any local wiki again but this time around, set `usesul3=0`, that
  should disable SUL3 and default to SUL2 mode.
* visit a local wiki and don't set the `usesul3` param, SUL3 mode should
  still be triggered because we're defaulting to `always`. Without the
  `always` in the config array, we should use SUL2.

Bug: T377924
Change-Id: I1902acb4a44199e5fbb76a9cec88c14de250ceb3