SUL3: Make SharedDomainHookHandler restrictions configurable
Allows special page etc. restrictions (some if which are about
special pages in another extension, and in the case of disallowed
providers might even be private mitigations) to be configured via
$wgCentralAuthSul3SharedDomainRestrictions, rather than hardcoded.

Allowing such restrictions to be defined in the extension owning
the special page etc. would be more elegant, but with CentralAuth
not really supported outside Wikimedia, didn't seem worth the
effort.

Also fixes a major bug: ApiCheckCanExecute does not work on
submodules.

Bug: T377142
Depends-On: I88effab746544c4788b607059a2184ead5b1bbac
Change-Id: Ie407c14e7a345f1b93537d08782f303e107ee1b2