SUL3: Disable more auth providers in the local leg of SUL3 login
Disable various authentication providers which do not normally
require user interaction, but can prevent authentication attempts
under the right circumstances. We want such prevention to happen
on the central wiki, not the local wiki (where the user would be
left with a confusing empty login page with just a "central login"
button).

See T369180#10098158 for a more detailed analysis of which
providers are being disabled.

Many of these are from random extensions, with the only shared
feature being that they are used in Wikimedia production. We now
have a way to set these in site configuration rather than
CentralAuth code, but given that CentralAuth is already highly
Wikimedia-specific, and splitting this list in two parts which
are in separate repositories with different deploy schedules
seems like a pain, it didn't seem worth it.

Note: deploying this will break in-progress login sessions, as
AuthManager does not like it when the list of available provider
keys changes mid-flight. There will be some logspam for up to five
minutes after deployment ("Authentication failed because of
inconsistent provider array"). That's unavoidable, and low-impact
(users will see a generic login error; their next login attempt
will start a new session and succeed).
To minimize such disruption, ideally it should be merged on the
same week as the patch it depends on.

Bug: T369180
Depends-On: I71427b2bb1890dc8c22844115dccbe8078a20568
Change-Id: I8e15478697de105ba4e2466648975741591b115d