Introduce MainConfigNames for CentralAuth config settings
This patch introduces a class for accessing CentralAuth config
settings for convenience. On advantage of such a class is to detect
unused config settings and remove them since they can't be noticed
from looking at our extension registration schema (extension.json)
file.

Another benefit is that when we use the @deprecated annotation on
a constant, it shows in all places referenced that it has been done
so, that way, it encourages developers via static analysis to use the
new setting name or cleanup their code. But with extension.json, it
is hard to know without reading the documentation that a setting has
been deprecated. Example is the CentralAuthDatabase config setting.

The patch also references config settings that are already in core,
making use of the MainConfigNames in core rather than type the name
of the setting by hand.

Change-Id: Iec971e35a3932f5c52a2e7bd520bf9db11f6e30a