Point autologin resourceloader module URL to `/start` endpoint
This patch makes CentralAuth's resourceloader `autologin` module that
is in charge of auto logins to use the `/start` endpoint instead of
`/checkLoggedIn` so that in the near future, accessing the `/checkedLoggedIn`
endpoint is no longer hardcoded to go via the central login wiki.

The above responsibility of accessing the central domain to check for
a valid logged-in session will be directed by the local wiki via the
`/start` endpoint.

The impact of this patch should not be noticeable by the end user and
SUL2 should still work the same way as before.

`@note`: For backwards compatibility, just retain the `checkLoggedIn`
URL for a week before removal.

Bug: T375788
Change-Id: I0f7828c957dcd6728aeda89a9caa705debff7e33