Re-enable `testViewForExistingGlobalTemporaryAccount`
This test has been causing issues in CI due to problematic
interactions with the CheckUser extension and Temporary User
functionality.

Since T378478, we run tests in order by filename. This should
reduce the cross-extension interactions. Any remaining issues
that show up as a result of this test being enabled should be
investigated and a root-cause for the state leak should be
found.

Bug: T377197
Change-Id: I26cd17bdd38bd217969a7ea30ffbe47b858a769b