Check session provider when autocreating
We allowed session as the autocreation source in testUserForCreation()
but ignored it in autoCreatedAccount() (which attaches autocreated
accounts to the local account). This resulted in unattached local
accounts being created when a centrally logged in user visited a new
wiki, and there already was an unattached account by the same name
elsewhere.

The purpose of the check was to prevent takeover of the global account
by autocreation from another provider. In practice CentralAuth isn't
used together with other providers that would autocreate, but still
nice to be defensive here, so invent a way do determine whether
autocreation-from-session is done by CentralAuth or another provider.
Unfortunately AuthManager does not provide a way to pass autocreation
source when autocreating from the session, and this didn't seem worth
changing the signature of a stable method; but it's easy to hack
around it.

Bug: T378289
Change-Id: Ic03c6d9171a090bc26d7eb51b460b576c6f27e23