Migrate installer sysop globalization to an installer task
This allows the installer to schedule sysop creation after extension
table creation. It elegantly ensures that the globalization happens only
in the web and CLI installer, not in update.php or
installPreConfigured.php.

The installer has a mode where it runs on a wiki with existing tables.
The existing code was potentially dangerous in that case, globalizing
all users on a wiki. So have core supply a list of usernames that the
installer actually created, and use that here to select users to
globalize.

Bug: T352113
Depends-On: I014a7accbedf66875115d188b38046c845482dc8
Change-Id: Ia3a94414f92574b3f1b267802874c4f4865972a5